test: run test-setproctitle where supported
Setting the process title has been enabled in libuv on AIX and z/OS. The
latest level of libuv skips only skips testing of uv_set_process_title
when __sun is #defined.

This change simplifies the skip test so the test is only skipped when
common.isSunOS is true to match libuv. Skip running the `ps` part of the
test on Windows.

PR-URL: https://github.com/nodejs/node/pull/11416
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Gibson Fahnestock <gibfahn@gmail.com>
Reviewed-By: Michael Dawson <michael_dawson@ca.ibm.com>